modularize models and update docs
