--- src/models/category.rb
@@ -1,6 +1,0 @@
-class Category < ApplicationRecord
-  has_many :products
-
-  # Validations
-  validates :name, presence: true, uniqueness: true
-end
