+++ src/models/products/category.rb
@@ -1,0 +1,9 @@
+module Products
+  class Category < ApplicationRecord
+    has_many :products
+  
+    # Validations
+    validates :name, presence: true, uniqueness: true
+  end
+  
+end
