--- src/models/user.rb
@@ -1,6 +1,0 @@
-class User < ApplicationRecord
-  has_many :orders
-  has_many :customizations
-
-  # validations go here
-end
