+++ src/models/users/user.rb
@@ -1,0 +1,8 @@
+module Users
+  class User < ApplicationRecord
+    has_many :orders
+    has_many :customizations
+  
+    # validations go here
+  end
+end
