+++ src/services/products/product_service.rb
@@ -3,10 +3,10 @@ module Products
     def self.create_with_default_config(category_name, product_details)
       ActiveRecord::Base.transaction do
   
-        category = Category.find_by(name: category_name)
+        category = Products::Category.find_by(name: category_name)
         product_config = Config::ConfigService.new(category).default_product_configuration
   
-        product = Product.create!(
+        product = Products::Product.create!(
           name: product_details[:name],
           description: product_details[:description],
           category: category
@@ -17,11 +17,11 @@ module Products
     end
   
     def self.create_part(part_details)
-      Part.create!(part_details)
+      Products::Part.create!(part_details)
     end
   
     def self.create_part_option(part_option_details)
-      PartOption.create!(
+      Products::PartOption.create!(
         name: part_option_details[:name],
         price: part_option_details[:price],
         in_stock: part_option_details[:in_stock],
@@ -30,15 +30,15 @@ module Products
     end
   
     def self.create_part_option_combinations(part_option_combination_details)
-      PartOptionCombination.create!(part_option_combination_details)
+      Products::PartOptionCombination.create!(part_option_combination_details)
     end
   
     def self.create_part_option_prohibitions(prohibition_details)
-      PartOptionProhibition.create!(prohibition_details)
+      Products::PartOptionProhibition.create!(prohibition_details)
     end
   
     def self.update_combination_price(new_combination_price_details)
-      PartOptionCombination.update(
+      Products::PartOptionCombination.update(
         new_combination_price_details[:id],
         combination_price: new_combination_price_details[:price_adjustment]
       )
@@ -47,7 +47,7 @@ module Products
     # Always override part_options value
     # Reason: UI reflects true state each time. Admin can unselect / select
     def self.update_prohibited_part_options(update_prohibition_details)
-      PartOptionProhibition.update(
+      Products::PartOptionProhibition.update(
         update_prohibition_details[:id],
         part_options: update_prohibition_details[:part_options]
       )
